Factor aggregation job handling to seperate state machine
